chore:migrated Favorite Restaurants screen into Jetpack compose

+++ app/src/main/java/com/sorabh/grabfood/ui/viewmodel/FavoriteRestaurantsViewModel.kt
@@ -5,11 +5,10 @@ import androidx.lifecycle.viewModelScope
 import com.sorabh.grabfood.domain.model.reataurants_home_response.Dish
 import com.sorabh.grabfood.domain.repository.LocalDBRepository
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.async
+import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
@@ -20,37 +19,56 @@ class FavoriteRestaurantsViewModel @Inject constructor(private val localDBReposi
     private val _getDishesFlow: MutableStateFlow<List<Dish>> = MutableStateFlow(emptyList())
     val getDishesFlow: StateFlow<List<Dish>> = _getDishesFlow
 
+//    val isFavoriteDish = MutableStateFlow(false)
+
     init {
         viewModelScope.launch {
             getDishes()
         }
     }
-    private suspend fun getDishes(){
-        localDBRepository.getRestaurantList().collect{
+
+    private suspend fun getDishes() {
+        localDBRepository.getRestaurantList().collect {
             _getDishesFlow.emit(it)
         }
     }
 
-    suspend fun insertRestaurant(dish:Dish):Boolean{
-        val result = CoroutineScope(Dispatchers.IO).async {
-            localDBRepository.insertRestaurant(dish)
-            isRestaurantStored(dish.id)
+    fun onFavoriteIconBtnClick(dish: Dish) {
+        viewModelScope.launch {
+            if (localDBRepository.getRestaurant(dish.id) > 0)
+                localDBRepository.deleteRestaurant(dish)
+            else
+                localDBRepository.insertRestaurant(dish)
         }
-        return result.await()
     }
 
-    suspend fun deleteRestaurant(dish:Dish):Boolean{
-        val result = CoroutineScope(Dispatchers.IO).async {
-            localDBRepository.deleteRestaurant(dish)
-            isRestaurantStored(dish.id)
-        }
-        return !result.await()
+    fun isDishStored(id: String): Flow<Boolean> = flow {
+        emit(
+            localDBRepository.getRestaurant(id) > 0
+        )
     }
+//
+//    suspend fun insertRestaurant(dish: Dish): Boolean {
+//        val result = viewModelScope.async {
+//            localDBRepository.insertRestaurant(dish)
+//            isRestaurantStored(dish.id)
+//        }
+//        return result.await()
+//    }
 
-    suspend fun isRestaurantStored(id:String):Boolean{
-        val result = CoroutineScope(Dispatchers.IO).async {
-            localDBRepository.getRestaurant(id)
-        }
-        return result.await() > 0
-    }
+//    suspend fun deleteRestaurant(dish: Dish): Boolean {
+//        val result = viewModelScope.async {
+//            localDBRepository.deleteRestaurant(dish)
+//            isRestaurantStored(dish.id)
+//        }
+//        return !result.await()
+//    }
+
+//    fun isRestaurantStored(id: String): Boolean {
+//        val result = viewModelScope.async {
+//            localDBRepository.getRestaurant(id)
+//        }
+//        isFavoriteDish.value = result.await() > 0
+//        return isFavoriteDish.value
+//    }
 }
